+++ motor-control/tests/test_motor_interrupt_handler.cpp
@@ -31,7 +31,8 @@ SCENARIO("a move is cancelled due to a stop request") {
                         .acceleration = 2,
                         .group_id = 0,
                         .seq_id = 0,
-                        .stop_condition = MoveStopCondition::limit_switch};
+                        .stop_condition = static_cast<uint8_t>(
+                            MoveStopCondition::limit_switch)};
         test_objs.queue.try_write_isr(msg);
         test_objs.queue.try_write_isr(msg);
         test_objs.queue.try_write_isr(msg);
@@ -76,7 +77,8 @@ SCENARIO("estop pressed during motor interrupt handler") {
                         .acceleration = 2,
                         .group_id = 0,
                         .seq_id = 0,
-                        .stop_condition = MoveStopCondition::limit_switch};
+                        .stop_condition = static_cast<uint8_t>(
+                            MoveStopCondition::limit_switch)};
         test_objs.queue.try_write_isr(msg);
         test_objs.queue.try_write_isr(msg);
         test_objs.queue.try_write_isr(msg);
@@ -127,7 +129,8 @@ SCENARIO("estop is steady-state pressed") {
                         .acceleration = 2,
                         .group_id = 0,
                         .seq_id = 0,
-                        .stop_condition = MoveStopCondition::limit_switch};
+                        .stop_condition = static_cast<uint8_t>(
+                            MoveStopCondition::limit_switch)};
         test_objs.queue.try_write_isr(msg);
         msg.message_index = 2;
         test_objs.queue.try_write_isr(msg);
